Added saga for the upload process
