+++ src/redux/sagas/_root.saga.js
@@ -3,6 +3,7 @@ import loginSaga from './login.saga';
 import registrationSaga from './registration.saga';
 import userSaga from './user.saga';
 import closetSaga from './closet.saga';
+import uploadSaga from './upload.saga.js';
 
 // rootSaga is the primary saga.
 // It bundles up all of the other sagas so our project can use them.
@@ -17,5 +18,6 @@ export default function* rootSaga() {
     registrationSaga(),
     userSaga(),
     closetSaga(),
+    uploadSaga(),
   ]);
 }
